Move getAsterism to AsterismService
